requirements.txt
Flask>=2.2
python-dotenv>=0.21
requests>=2.28
pytest>=7.0
google-generativeai>=0.3.0
